+++ tests/e2e/cucumber/steps/api.ts
@@ -171,7 +171,6 @@ Given(
 )
 
 Given('{string} creates the following project space using API', async function (this: World, stepUser: string, stepTable: DataTable): Promise<void> {
-    const spaceAdmin = this.usersEnvironment.getUser({ key: stepUser })
     const user = this.usersEnvironment.getUser({ key: stepUser })
     for (const space of stepTable.hashes()) {
       const spaceId = await api.graph.createSpace({ user, space })
