+++ tests/e2e/support/api/davSpaces/index.ts
@@ -1,1 +1,2 @@
-export { createFolder } from './spaces'
+export { createFolder, uploadFileInPersonalSpace } from './spaces'
+export { createFolder, createFolderInsideSpace, getFileId, uploadFileInsideSpace } from './spaces'
