+++ tests/e2e/support/api/graph/index.ts
@@ -6,6 +6,4 @@ export {
   deleteGroup,
   addUserToGroup
 } from './userManagement'
-export { getPersonalSpaceId } from './spaces'
-
-export { createSpace } from './space'
+export { getPersonalSpaceId , createSpace} from './spaces'
